REGRESSION(286456@main) [WebDriver] WebAutomationSessionProxy::evaluateJavaScriptFunction completionHandler should always be called https://bugs.webkit.org/show_bug.cgi?id=282983
Reviewed by NOBODY (OOPS!).

Ensure we call the completionHandler for a given callbackID to avoid the
operation from blocking.

The expectsImplicitCallbackArgument parameter is related to the JS
function being evaluated having a "resolver" callback that will be
called with the result of the evaluation, and is already handled by the
JS implementation of evaluateJavaScriptFunction.

* Source/WebKit/WebProcess/Automation/WebAutomationSessionProxy.cpp:
(WebKit::WebAutomationSessionProxy::evaluateJavaScriptFunction):