[WebDriver][BiDi] Implement the browsingContext.getTree method https://bugs.webkit.org/show_bug.cgi?id=283784
Reviewed by NOBODY (OOPS!).

Add support for the root and maxDepth parameter. The root selection is
done by splitting the steps of selecting the navigables to traverse
and actually traversing them.

Also, fix the presence of some fields in the returned payload. In the
current version of the spec, only the "parent" field is actually
optional, while the other fields must be present, even if they're
null.

In order to support the this, this commit also adds support for a new
boolean field in object types declared in the automation protocol:
"nullable". This field makes the generated C++ code accept optional
types (`RefPtr`, `std::optional`) in the setter, using JSON Null as the
underlying JSON value if the incoming value is not set.

This "nullable" tag contrasts with "optional" in the sense that with
optional, the key might be missing from the final object.

Note that this doesn't cover nullable command parameters nor support for
processing of incoming frontend data with nullable fields, as we only
need to support outbound data from the backend (i.e. command return
values) for now.

* Source/JavaScriptCore/inspector/scripts/codegen/cpp_generator.py:
(CppGenerator.cpp_type_for_type_member_argument):
* Source/JavaScriptCore/inspector/scripts/codegen/generate_cpp_protocol_types_header.py:
* Source/JavaScriptCore/inspector/scripts/codegen/models.py:
(Protocol.parse_type_member):
(TypeMember.__init__):
* Source/JavaScriptCore/inspector/scripts/tests/expected/type-declaration-object-type.json-result:
* Source/JavaScriptCore/inspector/scripts/tests/type-declaration-object-type.json:
* Source/WebKit/UIProcess/Automation/BidiBrowsingContextAgent.cpp:
(WebKit::BidiBrowsingContextAgent::getBrowsingContextID const):
(WebKit::BidiBrowsingContextAgent::getNavigableInfo):
(WebKit::BidiBrowsingContextAgent::getNextTree):
(WebKit::toOptionalUint):
(WebKit::BidiBrowsingContextAgent::getTree):
* Source/WebKit/UIProcess/Automation/BidiBrowsingContextAgent.h:
* Source/WebKit/UIProcess/Automation/WebAutomationSession.h:
* Source/WebKit/UIProcess/Automation/protocol/BidiBrowsingContext.json:
* WebDriverTests/TestExpectations.json: