[WebDriver][BiDi] Event timestamps should be an integral number https://bugs.webkit.org/show_bug.cgi?id=298656
Reviewed by NOBODY (OOPS!).

The spec states "time values" are represented as js-uint, which go from
0 to 2^53-1 (JS Max Safe integer). And, accordingly, client libraries
like the WPT tests and Selenium Python expect them to be integers.

On our side, currently we're using `number`/`double` for them, as the
"integer" primitive type is just a plain int. This works fine as long as
we ensure the fractional part is zero, so the JSON conversion drops it.

This commit truncates the timestamp values before passing them to the
inspector protocol code, as they'll likely still contain a fractional
part after the conversion to milliseconds.

In a follow-up commit, ideally we would add a primitive type to
the inspector protocol that would map to uint64_t to better enforce
this.

* Source/WebKit/UIProcess/Automation/WebAutomationSession.cpp:
(WebKit::WebAutomationSession::navigationStartedForFrame):
(WebKit::WebAutomationSession::navigationCommittedForFrame):
(WebKit::WebAutomationSession::navigationFailedForFrame):
(WebKit::WebAutomationSession::navigationAbortedForFrame):
(WebKit::WebAutomationSession::fragmentNavigatedForFrame):
(WebKit::WebAutomationSession::logEntryAdded):