[WebDriver][WPE] Take screenshots in the UIProcess
Currently, WPE's implentation of taking screenshots rely on asking the
root `RenderLayer` on the WebProcess to paint its content. While it
worked fine for most elements, this approach had limitations when
working with some features like CSS 3D filters and othe accelerated
content.

On top of that, a recent ANGLE bump included a stricter check on
ReadPixels that affected taking screenshots of WebGL scenes on WebGL 1.0
environments.

To address these issues, this commits adds support for taking
screenshots on the UIProcess, by reading the buffer that will be sent
to the system.

Currently, these are the kinds of WPEBuffer tested and supported:

DMA-BUF buffers (through WPE_DMABUF_BUFFER_FORMAT envvar):
- ARGB8888 or XRGB8888 pixel formats
- Linear or "Invalid" Buffer format modifiers

SHM buffers:
- ARGB8888 linear

When the view is static, the rendering pipeline might be dormant. We
need to ensure we'll receive an incoming buffer so we can read the
screenshot from it.

Note: This does not solve the problem of the compositing loop pausing
when the page is not visible. In this case, the screenshot operation
would still block indefinitively. Maybe we'd need some sort of flag to
keep it running for a simple cycle?